src/requirements.txt
Jetson.GPIO==2.0.17
rplcd==1.3.0
smbus2==0.4.1
websocket-client==1.2.1
